requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
torch==2.2.2
torchvision==0.17.2
opencv-python==4.9.0.80
pillow==10.4.0
numpy==1.26.4